Sampler
This change reimplements the sampler for the discrete Gaussian distribution used in all the schemes based on the LWE problem. It offers three ways of sampling: with negative sampling, with precomputed cumulative distribution, and with double sampling - an algorithm based on the ideas from the paper [1]. While sampling with precomputed cumulative distribution is the fastest, it is feasible only when the variance of the distribution is not too big. The double sampling is an alternative approach avoiding this problem with slightly slower performance. All the samplers need a precise computation of the exponential function, which could present a bottleneck for the algorithms. This is avoided using precomputed values. The exponential function is computed using Taylor polynomials.
Additionally, parameters for the LWE schemes are now automatically generated, although the correctness of this part still needs to be confirmed by the authors of the schemes due to some typos in their paper.

[1] https://eprint.iacr.org/2013/383