Changing ElGamal group generation and its usages (#6)
We correct the ElGamal group generation to give g a generator of the subgroup of quadratic residues of Z_p, where p is a strong prime. This change is done since it is known that DDH assumption does not hold in general Z_p* group while it is believed it holds in the subgroup of quadratic residues.