An ABE scheme (#10)
ABE schemes provide a functionality that an entity can decrypt a ciphertext if and only if he or she possesses sufficient set of attributes.
This PR implements a key policy (KP) ABE scheme presented in [this paper](https://eprint.iacr.org/2006/309.pdf). It is based on a monotone span programs (MSP) providing a policy for the decryption. It includes the following functionalities:
- encryption of a message with only specifying which set of attributes will be important for decryption policy
- a policy convertor that transforms a boolean formula expressing the policy which attributes are needed for the decryption into an equivalent MSP structure
- key derivation that given given a MSP structure creates a set of keys corresponding to attributes
- key delegation that delegates the above keys to entities with some attributes
- decryption that based on given keys decrypts or denies the user

As a part of the decryption process a helping function solving a matrix vector equation Ax = v is implemented using Gaussian elimination.