Dlog improvement (#12)
This pull request introduces a new version of the baby-step giant-step algorithm for calculating the discrete logarithm in Z_p and bn256 group. It is designed in a way that the algorithm prefers finding a smaller solution for a minimal additional computations. On average the computational complexity is comparable with the standard baby-step giant-step algorithm, in the worst case when the solution is close to the bound the complexity is less than doubled than the standard version, but for may practical cases when the bound is much greater than the expected solution it performs much faster. This is particularly useful for the quadratic and inner-product schemes, where the bound is put quite high due to the theoretical bounds.
This pull request also adds:
- parallel version for computing discrete log in bn256
- small changes in testing discrete logarithms avoiding the usage of library emmy